* ODE-890, apply Gary's fix, thanks Gary.
git-svn-id: https://svn.apache.org/repos/asf/ode/trunk@1000729 13f79535-47bb-0310-9956-ffa450edef68

+++ bpel-runtime/src/main/java/org/apache/ode/bpel/engine/BpelEngineImpl.java
@@ -793,7 +793,7 @@ public class BpelEngineImpl implements BpelEngine {
     }
 
     public BpelProcess getNewestProcessByType(QName processType) {
-        int v = -1;
+        long v = -1;
         BpelProcess q = null;
         for (BpelProcess p : _activeProcesses.values()) {
             if (p.getProcessType().equals(processType) && v < p.getVersion()) {
